niscope_acquire_waveform: Add _helpers.py
